adjust data synchronizer

+++ abremir.AllMyBricks.ThirdParty.Brickset/Models/LegoComDetails.cs
@@ -4,7 +4,7 @@ namespace abremir.AllMyBricks.ThirdParty.Brickset.Models
 {
     public class LegoComDetails
     {
-        public decimal? RetailPrice { get; set; }
+        public float? RetailPrice { get; set; }
         public DateTime? DateFirstAvailable { get; set; }
         public DateTime? DateLastAvailable { get; set; }
     }
